refactor: features/comment-management 분리
