+++ src/entities/post/ui/index.ts
@@ -1,1 +1,1 @@
-export { PostItem } from './PostItem';
+export * from './PostItem';
